requirements.txt
pandas
numpy
matplotlib
seaborn
scikit-learn
tensorflow
faiss-cpu
fastapi
uvicorn
streamlit
joblib
requests    